Moved utils to folder
