+++ OpenOrchestrator/database/db_utils/trigger_util.py
